Update: Refactor toast
