+++ src/shared/ui/components/toast-factory/index.ts
@@ -1,4 +1,3 @@
 export { useToastHook } from './use-toast-hook';
-export { ToastProvider } from './use-toast-context';
-export { ToastContext } from './toast-context';
+export { ToastProvider } from './toast-provider';
 export { useToast } from './toast-context';
